requirements.txt
gradio>=4.0.0
pillow>=9.3.0
python-dotenv>=1.0.0
requests>=2.28.1
google-generativeai>=0.3.0
torch>=2.0.0
numpy>=1.22.0
pydub>=0.25.1
coqui-tts
google-genai
soundfile
torch
torchvision
torchaudio